PostFiltering com estrategias de filterin + Amazon dataset partial

+++ src/br/cin/tbookmarks/recommender/database/contextual/AbstractContextualAttribute.java
@@ -1,5 +1,6 @@
 package br.cin.tbookmarks.recommender.database.contextual;
 
 public interface AbstractContextualAttribute {
-	public int getCode();
+		
+	public long getCode();
 }
